+++ ndnSimV24/ns-3/src/ndnSIM/helper/ndn-fib-helper.hpp
@@ -59,7 +59,7 @@ public:
   static void
   	AddRoute(Ptr<Node> node, const Name& prefix, shared_ptr<Face> face,
   			int32_t metric, std::string macAddress, double positionX,
-  			double positionY , double positionZ, double futurePositionX, double futurePositionY, double timeAtFuturePosition, int isFuturePositionSet);
+  			double positionY , double positionZ, double futurePositionX, double futurePositionY, double timeAtFuturePosition, double isFuturePositionSet);
 
   /**
    * \brief Add forwarding entry to FIB
@@ -77,7 +77,7 @@ public:
   static void
   	AddRoute(const std::string& nodeName, const Name& prefix, uint32_t faceId,
   			int32_t metric, std::string macAddress, double positionX,
-  			double positionY , double positionZ, double futurePositionX, double futurePositionY, double timeAtFuturePosition, int isFuturePositionSet);
+  			double positionY , double positionZ, double futurePositionX, double futurePositionY, double timeAtFuturePosition, double isFuturePositionSet);
 
   /**
    * \brief Add forwarding entry to FIB
@@ -96,7 +96,7 @@ public:
 	AddRoute(Ptr<Node> node, const Name& prefix, uint32_t faceId,
 			int32_t metric, std::string macAddress, double positionX,
 			double positionY, double positionZ, double futurePositionX,
-			double futurePositionY, double timeAtFuturePosition, int isFuturePositionSet);
+			double futurePositionY, double timeAtFuturePosition, double isFuturePositionSet);
 
   /**
    * @brief Add forwarding entry to FIB (work only with point-to-point links)
@@ -115,7 +115,7 @@ public:
   	AddRoute(Ptr<Node> node, const Name& prefix, Ptr<Node> otherNode,
   			int32_t metric, std::string macAddress, double positionX,
   			double positionY, double positionZ, double futurePositionX,
-  			double futurePositionY, double timeAtFuturePosition, int isFuturePositionSet);
+  			double futurePositionY, double timeAtFuturePosition, double isFuturePositionSet);
 
   /**
    * @brief Add forwarding entry to FIB (work only with point-to-point links)
@@ -137,7 +137,7 @@ public:
   			const std::string& otherNodeName, int32_t metric,
   			std::string macAddress, double positionX, double positionY,
   			double positionZ, double futurePositionX, double futurePositionY,
-  			double timeAtFuturePosition, int isFuturePositionSet);
+  			double timeAtFuturePosition, double isFuturePositionSet);
 
 
   /**
@@ -156,7 +156,7 @@ public:
 	RemoveRoute(Ptr<Node> node, const Name& prefix, shared_ptr<Face> face,
 			std::string macAddress, double positionX, double positionY,
 			double positionZ, double futurePositionX, double futurePositionY,
-			double timeAtFuturePosition, int isFuturePositionSet);
+			double timeAtFuturePosition, double isFuturePositionSet);
 
   /**
    * \brief remove forwarding entry in FIB
@@ -174,7 +174,7 @@ public:
 	RemoveRoute(Ptr<Node> node, const Name& prefix, uint32_t faceId,
 			std::string macAddress, double positionX, double positionY,
 			double positionZ, double futurePositionX, double futurePositionY,
-			double timeAtFuturePosition, int isFuturePositionSet);
+			double timeAtFuturePosition, double isFuturePositionSet);
 
   /**
    * \brief remove forwarding entry in FIB
@@ -192,7 +192,7 @@ public:
 	RemoveRoute(const std::string& nodeName, const Name& prefix,
 			uint32_t faceId, std::string macAddress, double positionX,
 			double positionY, double positionZ, double futurePositionX,
-			double futurePositionY, double timeAtFuturePosition, int isFuturePositionSet);
+			double futurePositionY, double timeAtFuturePosition, double isFuturePositionSet);
 
   /**
    * @brief remove forwarding entry in FIB (work only with point-to-point links)
@@ -210,7 +210,7 @@ public:
   	RemoveRoute(Ptr<Node> node, const Name& prefix, Ptr<Node> otherNode,
   			std::string macAddress, double positionX, double positionY,
   			double positionZ, double futurePositionX, double futurePositionY,
-  			double timeAtFuturePosition, int isFuturePositionSet);
+  			double timeAtFuturePosition, double isFuturePositionSet);
 
   /**
    * @brief remove forwarding entry in FIB (work only with point-to-point links)
@@ -229,7 +229,7 @@ public:
   			const std::string& otherNodeName, std::string macAddress,
   			double positionX, double positionY, double positionZ,
   			double futurePositionX, double futurePositionY,
-  			double timeAtFuturePosition, int isFuturePositionSet);
+  			double timeAtFuturePosition, double isFuturePositionSet);
 
 private:
   static void
